Puzzle 2 First Look
Created 2 types of trapstep on stairs in first floor. The first trapstep is when player steps on it, whole stairs shakes and the second trapstep type is when player steps on random step, player falls after little delay sec under stairs.
(total nbr of trapsteps are for now 8 covering whole stairs in first floor!)
So player here need to use the jumping feature that was implemented earlier in 0.6 in order to avoid stepping on trapsteps.

note: More sound effects and adjustments for shaking and falling will be added later in next commits and more than that to make puzzle 2 more cooler!!